FIX load dense model in randomly_prune.py
FIX in randomly_prune.py, open_lth previously loaded a checkpoint
with the same sparsity level.
This causes the random mask to be applied on top of the
original mask for this level, i.e. taking the intersection of the masks.
The fix is to load the dense model  from pretrain (same as one_shot.py)