ADD layernorm scaling hack
ADD layernorm_scaling argument for loading models through api.py.
The purpose of this argument is to rescale layernorm stats
to the correct size when models are padded to increase/decrease width
for partial permutation alignment.
This is a hack to accommodate existing layernorm checkpoints as a correction
using the affine parameters of layernorm is hard to compute.
Scaling is not needed during training because the models are only padded
at inference time.
This approach does not add a new hparam in order to avoid changing
existing experiment hashes.